// frontend/src/app/api/models/catalog-overview.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Assumptions } from '../models/assumptions';
import { CatalogModel } from '../models/catalog-model';
import { HardwareInfo } from '../models/hardware-info';
import { ModelGroup } from '../models/model-group';
import { OllamaState } from '../models/ollama-state';
import { RemovedModel } from '../models/removed-model';
import { ServerConfig } from '../models/server-config';
export interface CatalogOverview {
  asOf: string;
  assumptions: Assumptions;
  groups: Array<ModelGroup>;
  hardware: HardwareInfo;
  models: Array<CatalogModel>;
  ollama: OllamaState;
  refreshedAt?: (string | null);
  removed: Array<RemovedModel>;
  revision: number;
  server: ServerConfig;
}